config utilities

+++ main.sh
@@ -18,6 +18,8 @@ apt install -y firmware-amd-graphics
 apt install -y firmware-iwlwifi
 apt install -y firmware-atheros
 apt install -y firmware-realtek
+apt install -y firmware-b43legacy-installer
+apt install -y firmware-b43-installer
 apt install -y intel-microcode intel-gpu-tools
 apt install -y amd64-microcode amdgcn-tools
 #INTEL e AMD
@@ -41,7 +43,7 @@ apt install -y mousepad xarchiver
 apt install -y xdg-user-dirs xdg-user-dirs-gtk
 apt install -y elementary-xfce-icon-theme
 #Fonts Padrão
-
+#apt install -y fonts-noto* fonts-powerline fonts-firacode
 #Software Padrão
 apt install -y firefox-esr atril gnome-calculator
 #Gerenciador Login
@@ -56,7 +58,7 @@ apt install -y bash-completion command-not-found
 apt install -y zip p7zip* unrar* unzip rar arc arj cabextract lhasa unace* xz-utils sharutils uudeview mpack
 #Codecs de audio e video
 apt install -y ffmpeg mencoder faad lame sox twolame vorbis-tools libavcodec-extra* gstreamer1.0-fdkaac gstreamer1.0-plugins-ugly
-
+####Config####
 #Habilitar Theme Grub
 sed -i 's/#GRUB_GFXMODE=640x480/GRUB_GFXMODE=1024x768/g' /etc/default/grub
 sed -i 's/GRUB_CMDLINE_LINUX_DEFAULT="quiet"/GRUB_CMDLINE_LINUX_DEFAULT="quiet splash modprobe.blacklist=pcspkr"/g' /etc/default/grub
@@ -75,13 +77,16 @@ cp -f config/interfaces /etc/network/interfaces
 #Aplicar Thema do Boot e Atualizar Boot
 plymouth-set-default-theme -R bgrt
 update-grub2
-#Limpeza no apt
-apt autoremove -y
-apt autoclean
-apt clean
-
-source utilities.sh
-
-#Reinicia o sistema
-read -r -p "Instalação concluida! Seu pc precisa ser reiniciad! [Enter] " REBOOT
-reboot
+#Reiniciar ou Software Extra
+read -r -p "Deseta instalar os programas complementares? [y|n] " SOFTWARE
+if [ "$SOFTWARE" = "n" ]; then
+    #Limpeza no apt
+    apt autoremove -y
+    apt autoclean
+    apt clean
+    #Reinicia o sistema
+    read -r -p "Instalação concluida! O sistema será reinicializado! [Enter] " REBOOT
+    reboot
+else
+    sh utilities.sh
+fi
